Refactor static checking
Extracted the two phases of static checking into their own methods, and
added a method for just running the static checks on a list of files.

Refactored the tests now make use of this, which means that the logic we
test should be exactly the same as the one we run.